renamed mock location
