+++ libandroid-navigation-ui/src/main/java/com/mapbox/services/android/navigation/ui/v5/LocationViewModel.java
@@ -12,7 +12,7 @@ import android.preference.PreferenceManager;
 
 import com.mapbox.directions.v5.models.DirectionsRoute;
 import com.mapbox.mapboxsdk.location.LocationSource;
-import com.mapbox.services.android.navigation.v5.mock.MockLocationEngine;
+import com.mapbox.services.android.navigation.v5.location.MockLocationEngine;
 import com.mapbox.services.android.navigation.v5.navigation.NavigationConstants;
 import com.mapbox.services.android.navigation.v5.navigation.NavigationRoute;
 import com.mapbox.services.android.telemetry.location.LocationEngine;
