+++ libandroid-navigation-ui/src/main/java/com/mapbox/services/android/navigation/ui/v5/instruction/turnlane/TurnLaneAdapter.java
@@ -10,7 +10,6 @@ import com.mapbox.services.android.navigation.ui.v5.R;
 import com.mapbox.services.commons.utils.TextUtils;
 
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 
 import static com.mapbox.services.android.navigation.v5.navigation.NavigationConstants.STEP_MANEUVER_MODIFIER_LEFT;
